# docs/swagger.yaml
basePath: /api/v1
definitions:
  schema.AuthorizationSchema:
    properties:
      login_or_email:
        type: string
      password:
        type: string
    required:
    - login_or_email
    - password
    type: object
  schema.RegistrationSchema:
    properties:
      email:
        type: string
      full_name:
        type: string
      login:
        type: string
      password:
        minLength: 8
        type: string
    required:
    - email
    - full_name
    - login
    - password
    type: object
host: localhost:8080
info:
  contact: {}
  description: This is the API for the Tamagotchi application.
  termsOfService: http://localhost:8080/terms/
  title: Tamagotchi API
  version: "1"
paths:
  /user/auth:
    post:
      consumes:
      - application/json
      parameters:
      - description: AuthorizationSchema
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/schema.AuthorizationSchema'
      produces:
      - application/json
      responses:
        "400":
          description: Bad Request
          schema:
            type: object
        "500":
          description: Internal Server Error
          schema:
            type: object
      summary: Authorization
      tags:
      - users
  /user/logout:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "400":
          description: Bad Request
          schema:
            type: object
        "500":
          description: Internal Server Error
          schema:
            type: object
      summary: Logout
      tags:
      - users
  /user/registration:
    post:
      consumes:
      - application/json
      parameters:
      - description: RegistrationSchema
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/schema.RegistrationSchema'
      produces:
      - application/json
      responses:
        "400":
          description: Bad Request
          schema:
            type: object
        "500":
          description: Internal Server Error
          schema:
            type: object
      summary: Registration
      tags:
      - users
schemes:
- http
swagger: "2.0"